Flattening file structure
