--- lib/db.php
@@ -1,29 +1,0 @@
-<?php
-/*
-    Copyright 2010 Pieter Iserbyt
-
-    This file is part of Pamela.
-
-    Pamela is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Pamela is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Pamela.  If not, see <http://www.gnu.org/licenses/>.
-*/
-require_once("config.php");
-
-function get_db() {
-	static $db = NULL;
-	if ($db == NULL) {
-		$db = new SQLite3(SQLITE_DB);
-	}
-	return $db;
-}
-
